requirements.txt
python-dotenv==1.0.1
python-telegram-bot==21.3
Requests==2.32.3
worldnewsapi==1.2.0
